ita-stf: move module to seperate file
